Move KeyId primitive up to top namespace
