+++ tests/Tests/Isolated/Encryption/MessageTest.php
@@ -15,7 +15,7 @@ namespace OpenEMR\Tests\Isolated\Encryption;
 use BadMethodCallException;
 use OpenEMR\Encryption\{
     Ciphertext,
-    Keys\Id,
+    KeyId,
     Message,
     MessageFormat,
 };
@@ -82,7 +82,7 @@ class MessageTest extends TestCase
 
     public function testConstructWithImplicitKeyFormat(): void
     {
-        $keyId = new Id('005');
+        $keyId = new KeyId('005');
         $ciphertext = new Ciphertext('test data');
         $message = new Message($keyId, $ciphertext, MessageFormat::ImplicitKey);
 
@@ -92,7 +92,7 @@ class MessageTest extends TestCase
 
     public function testConstructWithInvalidImplicitKey(): void
     {
-        $keyId = new Id('not-a-number');
+        $keyId = new KeyId('not-a-number');
         $ciphertext = new Ciphertext('test data');
         $this->expectException(BadMethodCallException::class);
         $message = new Message($keyId, $ciphertext, MessageFormat::ImplicitKey);
@@ -101,7 +101,7 @@ class MessageTest extends TestCase
 
     public function testConstructParseRoundtripImplicitKey(): void
     {
-        $keyId = new Id('007');
+        $keyId = new KeyId('007');
         $ciphertext = new Ciphertext('some encrypted data');
         $message = new Message($keyId, $ciphertext, MessageFormat::ImplicitKey);
         $encoded = $message->encode();
